ellibcolorcorrect: fix reading auto location enabled from config
Reading the 'Mode' config key into an int instead of string returned 0, which was interpreted as auto location mode, and resulted in the locator always being started even when the user had not chosen auto location.

Also explicitly check for the NightColor Active config, to future proof it in case colorcorrect gets other uses such as adjusting screen gamma that can be running independently of Night Light being enabled.

BUG: 485086

FIXED-IN: 6.0.4